+++ plugin/v212/src/main/kotlin/dev/azn9/plugins/discord/postLoad/RenderAppLifecycleListener.kt
@@ -21,6 +21,7 @@ import dev.azn9.plugins.discord.render.renderService
 import dev.azn9.plugins.discord.utils.DisposableCoroutineScope
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.StartupActivity
+import dev.azn9.plugins.discord.DiscordPlugin
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.launch
@@ -30,6 +31,10 @@ class RenderPreloadingActivity : StartupActivity.Background, StartupActivity.Dum
 
     override fun runActivity(project: Project) {
         launch {
+            if (DiscordPlugin.isAlmightyAlpacasPluginPresent()) {
+                return@launch
+            }
+
             renderService.startRenderClock()
         }
     }