+++ plugin/v223/src/main/kotlin/dev/azn9/plugins/discord/time/TimeServiceAbstract.kt
@@ -107,10 +107,14 @@ class TimeService : TimeServiceAbstract, Disposable {
     }
 
     override fun load() {
+        if (DiscordPlugin.isAlmightyAlpacasPluginPresent()) {
+            return
+        }
+
         DiscordPlugin.LOG.debug("Loading TimeService vPRE-231")
 
-        if (loaded.getAndSet(true)) {
-            unload()
+        if (loaded.get()) {
+            return
         }
 
         initializeApplication(ApplicationManager.getApplication())
@@ -126,6 +130,7 @@ class TimeService : TimeServiceAbstract, Disposable {
         val timeoutMillis = settings.timeoutMinutes.getStoredValue() * 60 * 1000
         IdeEventQueue.invokeLater {
             IdeEventQueue.getInstance().addIdleListener(idleListener, timeoutMillis)
+            loaded.set(true)
         }
     }